requirements.txt
# Photoye - 本地智能照片管理助手
# 项目依赖配置文件
# 版本: 1.0 (阶段2)

# ============================================================================
# 阶段0 - 基础依赖 (环境搭建与地基工程)
# ============================================================================

# GUI框架
PyQt6>=6.4.0

# 科学计算基础
numpy>=1.21.0

# 图像处理基础
Pillow>=9.0.0

# ============================================================================
# 阶段2 - AI引擎依赖
# ============================================================================

onnxruntime>=1.15.0        # AI模型推理引擎
opencv-contrib-python>=4.7.0  # 包含 YuNet / SFace / FaceDetectorYN
dlib>=19.24.2             # Dlib 检测与识别模型支持（Windows 可使用预编译轮子）
tokenizers>=0.13.0        # OpenCLIP 文本编码 (本地 tokenizer.json )

# 阶段3 - 数据科学和聚类
# scikit-learn>=1.2.0        # 机器学习和聚类算法

# 阶段后期 - 可能的扩展依赖
# faiss-cpu>=1.7.0           # 高效相似度搜索(CPU版本)
# faiss-gpu>=1.7.0           # 高效相似度搜索(GPU版本)

# 开发和测试工具
# pytest>=7.0.0              # 单元测试框架
# black>=22.0.0               # 代码格式化工具
# flake8>=4.0.0               # 代码风格检查

# 打包和分发
# PyInstaller>=5.0.0          # 应用程序打包工具

# ============================================================================
# 安装说明
# ============================================================================
# 
# 阶段0安装命令:
# pip install -r requirements.txt
#
# 创建虚拟环境 (推荐):
# python -m venv photoye
# source photoye/bin/activate  # Windows: photoye\Scripts\activate
# pip install -r requirements.txt
#
# 升级pip (如果需要):
# pip install --upgrade pip
#